# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Evolvus Swagger API",
    "description": "Evolvus Spec for Swagger",
    "termsOfService": "http://swagger.io/terms/"
  },
  "basePath": "/",
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "api/role": {
      "get": {
        "tags": [
          "Role API"
        ],
        "summary": "Get a role from the database.",
        "description": "",
        "operationId": "getRole",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-TENANT-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-USER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-IPHEADER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ENTITY-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ACCESSLEVEL",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "post": {
        "tags": [
          "Role API"
        ],
        "summary": "Add a new role to the database",
        "description": "",
        "operationId": "addRole",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-TENANT-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-USER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-IPHEADER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ENTITY-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ACCESSLEVEL",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "Role object that needs to be added to the database",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Role"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/role/{roleName}": {
      "put": {
        "tags": [
          "Role API"
        ],
        "summary": "Update a role to the database",
        "description": "",
        "operationId": "updaterole",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Role object that needs to be updated to the database",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Role"
            }
          },
          {
            "in": "header",
            "name": "X-TENANT-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-USER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-IPHEADER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ENTITY-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ACCESSLEVEL",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/private/api/role/{id}": {
      "put": {
        "tags": [
          "Role API"
        ],
        "summary": "Update a role to the database",
        "description": "",
        "operationId": "updaterole",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Role object that needs to be updated to the database",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Role"
            }
          },
          {
            "in": "header",
            "name": "X-TENANT-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-USER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-IPHEADER",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ENTITY-ID",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "X-ACCESSLEVEL",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Role": {
      "type": "object",
      "required": [
        "tenantId",
        "roleType",
        "applicationCode",
        "roleName",
        "activationStatus",
        "associatedUsers",
        "menuGroup"
      ],
      "tags": [
        {
          "tenantId": "string",
          "roleType": "string",
          "txtType": "items",
          "applicationCode": "string",
          "roleName": "string",
          "associatedUsers": "number",
          "entityId": "string",
          "menuGroup": "items"
        }
      ],
      "properties": {
        "tenantId": {
          "type": "string",
          "example": "T001"
        },
        "roleType": {
          "type": "string",
          "example": "MAKER"
        },
        "txtType": {
          "type": "items",
          "example": [
            "UPDATE_ROLE",
            "CREATE_ROLE"
          ]
        },
        "applicationCode": {
          "type": "string",
          "example": "SANDSTORM"
        },
        "roleName": {
          "type": "string",
          "example": "CONSOLE_DEVELOPER"
        },
        "associatedUsers": {
          "type": "number",
          "example": "5"
        },
        "entityId": {
          "type": "string",
          "example": "H001B001"
        },
        "menuGroup": {
          "type": "items",
          "example": [
            {
              "menuItems": [
                {
                  "menuItemType": "button",
                  "applicationCode": "FLUX-CDA",
                  "menuItemCode": "ROLE",
                  "title": "Role Management",
                  "menuItemOrder": 1
                },
                {
                  "menuItemType": "menu",
                  "applicationCode": "FLUX-CDA",
                  "menuItemCode": "ROLE",
                  "title": "User Management",
                  "menuItemOrder": 2
                }
              ],
              "deletedFlag": 0,
              "menuGroupCode": "Administration",
              "title": "ADMINISTRATION",
              "applicationCode": "FLUX-CDA",
              "tenantId": "IVL",
              "createdBy": "pavithra",
              "menuGroupOrder": 1,
              "description": "ops role",
              "activationStatus": "ACTIVE",
              "processingStatus": "PENDING_AUTHORIZATION",
              "associatedUsers": 5,
              "createdDate": "2018-07-06T12:56:34.623Z"
            }
          ]
        }
      }
    },
...
  }
}